Add MultiModalEncoder interface
This interface is intended to replace SoundEncoder and TextEncoder as separate interfaces, but as a starting point, classes are provided to wrap existing implementations with the new interface.

The new interface will expose only a batch encoder interface (which an implementation can unbatch as necessary).

Adds a SoundOrTextEncoder to pair a Sound/Text encoder as in retrieval use cases.

PiperOrigin-RevId: 807765879